requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1.1
Django==3.1.1
django-admin-interface==0.13.5
django-celery-results==1.2.1
django-colorfield==0.3.2
django-cors-headers==3.5.0
django-crontab==0.7.1
django-environ==0.4.5
django-extra-fields==3.0.2
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-rest-knox==4.1.0
django-sslserver==0.22
djangorestframework==3.12.1
drf-yasg==1.17.1
httpagentparser==1.9.0
idna==2.10
importlib-metadata==2.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
mysqlclient==1.4.6
nameparser==1.0.6
numpy==1.19.2
packaging==20.4
pandas==1.1.2
Pillow==7.2.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
XlsxWriter==1.3.6
zipp==3.2.0
